Cleanup Attachment in FileUploadTestFactory
