+++ src/test/java/de/tum/in/www1/artemis/exercise/fileupload/FileUploadTestFactory.java
@@ -1,66 +1,19 @@
 package de.tum.in.www1.artemis.exercise.fileupload;
 
 import static java.time.ZonedDateTime.now;
-import static org.assertj.core.api.Assertions.fail;
 
-import java.io.File;
-import java.io.IOException;
-import java.nio.file.Path;
 import java.time.ZonedDateTime;
-import java.util.UUID;
 
-import org.apache.commons.io.FileUtils;
-import org.springframework.util.ResourceUtils;
-
-import de.tum.in.www1.artemis.domain.Attachment;
 import de.tum.in.www1.artemis.domain.Course;
 import de.tum.in.www1.artemis.domain.FileUploadExercise;
 import de.tum.in.www1.artemis.domain.FileUploadSubmission;
-import de.tum.in.www1.artemis.domain.enumeration.AttachmentType;
 import de.tum.in.www1.artemis.domain.exam.ExerciseGroup;
 import de.tum.in.www1.artemis.exercise.ExerciseTestFactory;
-import de.tum.in.www1.artemis.service.FilePathService;
-import de.tum.in.www1.artemis.service.FileService;
 
+/**
+ * Factory to create file upload exercises and submissions for testing.
+ */
 public class FileUploadTestFactory {
-
-    /**
-     * Creates a dummy attachment for testing.
-     *
-     * @param date The optional upload and release date to set on the attachment
-     * @return Attachment that was created
-     */
-    public static Attachment generateAttachment(ZonedDateTime date) {
-        Attachment attachment = new Attachment();
-        attachment.setAttachmentType(AttachmentType.FILE);
-        if (date != null) {
-            attachment.setReleaseDate(date);
-            attachment.setUploadDate(date);
-        }
-        attachment.setName("TestAttachment");
-        attachment.setVersion(1);
-        return attachment;
-    }
-
-    /**
-     * Creates a dummy attachment for testing with a placeholder image file on disk.
-     *
-     * @param startDate The release date to set on the attachment
-     * @return Attachment that was created with its link set to a testing file on disk
-     */
-    public static Attachment generateAttachmentWithFile(ZonedDateTime startDate) {
-        Attachment attachment = generateAttachment(startDate);
-        String testFileName = "test_" + UUID.randomUUID().toString().substring(0, 8) + ".jpg";
-        try {
-            FileUtils.copyFile(ResourceUtils.getFile("classpath:test-data/attachment/placeholder.jpg"), new File(FilePathService.getTempFilePath(), testFileName));
-        }
-        catch (IOException ex) {
-            fail("Failed while copying test attachment files", ex);
-        }
-        // Path.toString() uses platform dependant path separators. Since we want to use this as a URL later, we need to replace \ with /.
-        attachment.setLink(Path.of(FileService.DEFAULT_FILE_SUBPATH, testFileName).toString().replace('\\', '/'));
-        return attachment;
-    }
 
     /**
      * Creates a dummy file upload exercise for a course.
